Refactor project opening/saving into ProjectsStorage with StorageProviders.
