+++ newIDE/app/scripts/update-fixtures-from-resources-examples.js
@@ -36,15 +36,16 @@ const updateResources = (project, baseUrl) => {
   project.exposeResources(worker);
 };
 
-const writeBrowserExampleFilesJsFile = exampleNames => {
+const writeInternalExampleFilesJsFile = exampleNames => {
   let importsCode = [];
   let internalFilesObjectCode = [];
   exampleNames.forEach((exampleName, index) => {
-    importsCode.push(`import exampleFile${index} from '../fixtures/${exampleName}/${exampleName}.json';`);
+    importsCode.push(`import exampleFile${index} from '../../fixtures/${exampleName}/${exampleName}.json';`);
     internalFilesObjectCode.push(`  'example://${exampleName}': exampleFile${index},`);
   });
 
   const content = [
+    `// @flow`,
     `// This file is generated by update-fixtures-from-resources-examples.js script`,
     ``,
     importsCode.join('\n'),
@@ -58,7 +59,7 @@ const writeBrowserExampleFilesJsFile = exampleNames => {
 
   return new Promise((resolve, reject) => {
     fs.writeFile(
-      '../src/ProjectsStorage/BrowserExampleFiles.js',
+      '../src/ProjectsStorage/InternalFileStorageProvider/InternalExampleFiles.js',
       content,
       'utf8',
       err => {
@@ -110,14 +111,14 @@ extensionsLoader
           });
       })
     )
-      .then(() => writeBrowserExampleFilesJsFile(exampleNames))
+      .then(() => writeInternalExampleFilesJsFile(exampleNames))
       .then(
         () => {
-          console.error(`✅ BrowserExampleFiles.js written.`);
+          console.error(`✅ InternalExampleFiles.js written.`);
         },
         error => {
           console.error(
-            `❌ Error caught while writing BrowserExampleFiles.js:`,
+            `❌ Error caught while writing InternalExampleFiles.js:`,
             error
           );
         }
