+++ newIDE/app/src/BrowserApp.js
@@ -1,5 +1,5 @@
 // @flow
-import React from 'react';
+import * as React from 'react';
 import MainFrame from './MainFrame';
 import Window from './Utils/Window';
 import ExportDialog from './Export/ExportDialog';
@@ -10,8 +10,6 @@ import './UI/iconmoon-font.css'; // Styles for Iconmoon font.
 // Import for browser only IDE
 import BrowserExamples from './ProjectCreation/BrowserExamples';
 import BrowserStarters from './ProjectCreation/BrowserStarters';
-import BrowserProjectOpener from './ProjectsStorage/BrowserProjectOpener';
-import BrowserSaveDialog from './ProjectsStorage/BrowserSaveDialog';
 import BrowserIntroDialog from './MainFrame/BrowserIntroDialog';
 import browserResourceSources from './ResourcesList/BrowserResourceSources';
 import browserResourceExternalEditors from './ResourcesList/BrowserResourceExternalEditors';
@@ -22,6 +20,8 @@ import ObjectsEditorService from './ObjectEditor/ObjectsEditorService';
 import ObjectsRenderingService from './ObjectsRendering/ObjectsRenderingService';
 import { makeBrowserS3EventsFunctionCodeWriter } from './EventsFunctionsExtensionsLoader/CodeWriters/BrowserS3EventsFunctionCodeWriter';
 import Providers from './MainFrame/Providers';
+import ProjectsStorage from './ProjectsStorage';
+import InternalFileStorageProvider from './ProjectsStorage/InternalFileStorageProvider';
 
 export const create = (authentification: Authentification) => {
   Window.setUpContextMenu();
@@ -38,30 +38,38 @@ export const create = (authentification: Authentification) => {
       eventsFunctionsExtensionOpener={null}
     >
       {({ i18n, eventsFunctionsExtensionsState }) => (
-        <MainFrame
-          i18n={i18n}
-          eventsFunctionsExtensionsState={eventsFunctionsExtensionsState}
-          previewLauncher={<BrowserS3PreviewLauncher />}
-          renderExportDialog={(props) => <ExportDialog {...props} exporters={getBrowserExporters()} />}
-          createDialog={
-            <CreateProjectDialog
-              examplesComponent={BrowserExamples}
-              startersComponent={BrowserStarters}
+        <ProjectsStorage
+          storageProviders={[InternalFileStorageProvider]}
+          defaultStorageProvider={InternalFileStorageProvider}
+        >
+          {projectsStorageProps => (
+            <MainFrame
+              i18n={i18n}
+              eventsFunctionsExtensionsState={eventsFunctionsExtensionsState}
+              previewLauncher={<BrowserS3PreviewLauncher />}
+              renderExportDialog={props => (
+                <ExportDialog {...props} exporters={getBrowserExporters()} />
+              )}
+              createDialog={
+                <CreateProjectDialog
+                  examplesComponent={BrowserExamples}
+                  startersComponent={BrowserStarters}
+                />
+              }
+              introDialog={<BrowserIntroDialog />}
+              projectsStorage={projectsStorageProps}
+              resourceSources={browserResourceSources}
+              resourceExternalEditors={browserResourceExternalEditors}
+              authentification={authentification}
+              extensionsLoader={makeExtensionsLoader({
+                objectsEditorService: ObjectsEditorService,
+                objectsRenderingService: ObjectsRenderingService,
+                filterExamples: !Window.isDev(),
+              })}
+              initialPathsOrURLsToOpen={appArguments['_']}
             />
-          }
-          introDialog={<BrowserIntroDialog />}
-          saveDialog={<BrowserSaveDialog />}
-          onReadFromPathOrURL={BrowserProjectOpener.readInternalFile}
-          resourceSources={browserResourceSources}
-          resourceExternalEditors={browserResourceExternalEditors}
-          authentification={authentification}
-          extensionsLoader={makeExtensionsLoader({
-            objectsEditorService: ObjectsEditorService,
-            objectsRenderingService: ObjectsRenderingService,
-            filterExamples: !Window.isDev(),
-          })}
-          initialPathsOrURLsToOpen={appArguments['_']}
-        />
+          )}
+        </ProjectsStorage>
       )}
     </Providers>
   );
