+++ newIDE/app/src/LocalApp.js
@@ -13,8 +13,6 @@ import LocalExamples from './ProjectCreation/LocalExamples';
 import LocalStarters from './ProjectCreation/LocalStarters';
 import localResourceSources from './ResourcesList/LocalResourceSources';
 import localResourceExternalEditors from './ResourcesList/LocalResourceExternalEditors';
-import LocalProjectWriter from './ProjectsStorage/LocalProjectWriter';
-import LocalProjectOpener from './ProjectsStorage/LocalProjectOpener';
 import LocalPreviewLauncher from './Export/LocalExporters/LocalPreviewLauncher';
 import { getLocalExporters } from './Export/LocalExporters';
 import ElectronMainMenu from './MainFrame/ElectronMainMenu';
@@ -25,6 +23,8 @@ import ObjectsRenderingService from './ObjectsRendering/ObjectsRenderingService'
 import Providers from './MainFrame/Providers';
 import LocalEventsFunctionsExtensionWriter from './EventsFunctionsExtensionsLoader/Storage/LocalEventsFunctionsExtensionWriter';
 import LocalEventsFunctionsExtensionOpener from './EventsFunctionsExtensionsLoader/Storage/LocalEventsFunctionsExtensionOpener';
+import ProjectsStorage from './ProjectsStorage';
+import LocalFileStorageProvider from './ProjectsStorage/LocalFileStorageProvider';
 const gd = global.gd;
 
 export const create = (authentification: Authentification) => {
@@ -43,24 +43,28 @@ export const create = (authentification: Authentification) => {
         eventsFunctionsExtensionOpener={null}
       >
         {({ i18n, eventsFunctionsExtensionsState }) => (
-          <ExternalEditor
-            serverPort={appArguments['server-port']}
-            isIntegrated={appArguments['mode'] === 'integrated'}
-            editor={appArguments['editor']}
-            editedElementName={appArguments['edited-element-name']}
-          >
-            <MainFrame
-              i18n={i18n}
-              eventsFunctionsExtensionsState={eventsFunctionsExtensionsState}
-              resourceSources={localResourceSources}
-              authentification={authentification}
-              onReadFromPathOrURL={() =>
-                Promise.reject('Should never be called')
-              }
-              resourceExternalEditors={localResourceExternalEditors}
-              initialPathsOrURLsToOpen={[]}
-            />
-          </ExternalEditor>
+          <ProjectsStorage storageProviders={[]}>
+            {projectsStorageProps => (
+              <ExternalEditor
+                serverPort={appArguments['server-port']}
+                isIntegrated={appArguments['mode'] === 'integrated'}
+                editor={appArguments['editor']}
+                editedElementName={appArguments['edited-element-name']}
+              >
+                <MainFrame
+                  i18n={i18n}
+                  eventsFunctionsExtensionsState={
+                    eventsFunctionsExtensionsState
+                  }
+                  resourceSources={localResourceSources}
+                  authentification={authentification}
+                  projectsStorage={projectsStorageProps}
+                  resourceExternalEditors={localResourceExternalEditors}
+                  initialPathsOrURLsToOpen={[]}
+                />
+              </ExternalEditor>
+            )}
+          </ProjectsStorage>
         )}
       </Providers>
     );
@@ -74,36 +78,42 @@ export const create = (authentification: Authentification) => {
         eventsFunctionsExtensionOpener={LocalEventsFunctionsExtensionOpener}
       >
         {({ i18n, eventsFunctionsExtensionsState }) => (
-          <ElectronMainMenu i18n={i18n}>
-            <MainFrame
-              i18n={i18n}
-              eventsFunctionsExtensionsState={eventsFunctionsExtensionsState}
-              previewLauncher={<LocalPreviewLauncher />}
-              renderExportDialog={(props) => <ExportDialog {...props} exporters={getLocalExporters()} />}
-              createDialog={
-                <CreateProjectDialog
-                  examplesComponent={LocalExamples}
-                  startersComponent={LocalStarters}
+          <ProjectsStorage
+            storageProviders={[LocalFileStorageProvider]}
+            defaultStorageProvider={LocalFileStorageProvider}
+          >
+            {projectsStorageProps => (
+              <ElectronMainMenu i18n={i18n}>
+                <MainFrame
+                  i18n={i18n}
+                  eventsFunctionsExtensionsState={
+                    eventsFunctionsExtensionsState
+                  }
+                  previewLauncher={<LocalPreviewLauncher />}
+                  renderExportDialog={props => (
+                    <ExportDialog {...props} exporters={getLocalExporters()} />
+                  )}
+                  createDialog={
+                    <CreateProjectDialog
+                      examplesComponent={LocalExamples}
+                      startersComponent={LocalStarters}
+                    />
+                  }
+                  projectsStorage={projectsStorageProps}
+                  resourceSources={localResourceSources}
+                  resourceExternalEditors={localResourceExternalEditors}
+                  authentification={authentification}
+                  extensionsLoader={makeExtensionsLoader({
+                    gd,
+                    objectsEditorService: ObjectsEditorService,
+                    objectsRenderingService: ObjectsRenderingService,
+                    filterExamples: !Window.isDev(),
+                  })}
+                  initialPathsOrURLsToOpen={appArguments['_']}
                 />
-              }
-              onSaveProject={LocalProjectWriter.saveProject}
-              onSaveProjectAs={LocalProjectWriter.saveProjectAs}
-              onAutoSaveProject={LocalProjectWriter.autoSaveProject}
-              onChooseProject={LocalProjectOpener.chooseProjectFile}
-              onReadFromPathOrURL={LocalProjectOpener.readProjectFile}
-              shouldOpenAutosave={LocalProjectOpener.shouldOpenAutosave}
-              resourceSources={localResourceSources}
-              resourceExternalEditors={localResourceExternalEditors}
-              authentification={authentification}
-              extensionsLoader={makeExtensionsLoader({
-                gd,
-                objectsEditorService: ObjectsEditorService,
-                objectsRenderingService: ObjectsRenderingService,
-                filterExamples: !Window.isDev(),
-              })}
-              initialPathsOrURLsToOpen={appArguments['_']}
-            />
-          </ElectronMainMenu>
+              </ElectronMainMenu>
+            )}
+          </ProjectsStorage>
         )}
       </Providers>
     );
