--- newIDE/app/src/ProjectsStorage/BrowserExampleFiles.js
@@ -1,220 +1,0 @@
-// This file is generated by update-fixtures-from-resources-examples.js script
-
-import exampleFile0 from '../fixtures/admob/admob.json';
-import exampleFile1 from '../fixtures/advanced-shape-based-painter/advanced-shape-based-painter.json';
-import exampleFile2 from '../fixtures/animation-speed-scale/animation-speed-scale.json';
-import exampleFile3 from '../fixtures/asteroids/asteroids.json';
-import exampleFile4 from '../fixtures/basic-ai-with-pathfinding/basic-ai-with-pathfinding.json';
-import exampleFile5 from '../fixtures/basic-artificial-intelligence/basic-artificial-intelligence.json';
-import exampleFile6 from '../fixtures/basic-topdown-car-driving/basic-topdown-car-driving.json';
-import exampleFile7 from '../fixtures/betabox-basics-learning-experience/betabox-basics-learning-experience.json';
-import exampleFile8 from '../fixtures/bomb-the-crate/bomb-the-crate.json';
-import exampleFile9 from '../fixtures/bouncing-ball-and-rope/bouncing-ball-and-rope.json';
-import exampleFile10 from '../fixtures/breakout/breakout.json';
-import exampleFile11 from '../fixtures/buttons/buttons.json';
-import exampleFile12 from '../fixtures/car-physics/car-physics.json';
-import exampleFile13 from '../fixtures/center-object-within-another/center-object-within-another.json';
-import exampleFile14 from '../fixtures/change-position-of-object/change-position-of-object.json';
-import exampleFile15 from '../fixtures/change-scale-of-sprites/change-scale-of-sprites.json';
-import exampleFile16 from '../fixtures/change-sprite-animation/change-sprite-animation.json';
-import exampleFile17 from '../fixtures/change-sprite-color/change-sprite-color.json';
-import exampleFile18 from '../fixtures/character-selection/character-selection.json';
-import exampleFile19 from '../fixtures/create-object-with-mouseclick/create-object-with-mouseclick.json';
-import exampleFile20 from '../fixtures/custom-font/custom-font.json';
-import exampleFile21 from '../fixtures/custom-mouse-pointer/custom-mouse-pointer.json';
-import exampleFile22 from '../fixtures/customize-keys-with-lastpressedkey/customize-keys-with-lastpressedkey.json';
-import exampleFile23 from '../fixtures/device-orientation-ballgame/device-orientation-ballgame.json';
-import exampleFile24 from '../fixtures/device-orientation-compass/device-orientation-compass.json';
-import exampleFile25 from '../fixtures/device-vibration/device-vibration.json';
-import exampleFile26 from '../fixtures/dialogue-tree-with-yarn/dialogue-tree-with-yarn.json';
-import exampleFile27 from '../fixtures/downhill-bike-physics-demo/downhill-bike-physics-demo.json';
-import exampleFile28 from '../fixtures/drag-camera-with-mouse/drag-camera-with-mouse.json';
-import exampleFile29 from '../fixtures/drop-collect-items-from-storage/drop-collect-items-from-storage.json';
-import exampleFile30 from '../fixtures/endless-up-runner/endless-up-runner.json';
-import exampleFile31 from '../fixtures/exit-app/exit-app.json';
-import exampleFile32 from '../fixtures/facebook-instant-game/facebook-instant-game.json';
-import exampleFile33 from '../fixtures/filesystem-create-directory/filesystem-create-directory.json';
-import exampleFile34 from '../fixtures/find-diagonals/find-diagonals.json';
-import exampleFile35 from '../fixtures/geodash/geodash.json';
-import exampleFile36 from '../fixtures/health-bar/health-bar.json';
-import exampleFile37 from '../fixtures/infinite-scrolling-background/infinite-scrolling-background.json';
-import exampleFile38 from '../fixtures/inventory-system/inventory-system.json';
-import exampleFile39 from '../fixtures/isometric-game/isometric-game.json';
-import exampleFile40 from '../fixtures/javascript-blocks-in-platformer/javascript-blocks-in-platformer.json';
-import exampleFile41 from '../fixtures/keyboard-practice/keyboard-practice.json';
-import exampleFile42 from '../fixtures/level-editor/level-editor.json';
-import exampleFile43 from '../fixtures/load-image-from-url/load-image-from-url.json';
-import exampleFile44 from '../fixtures/magnet/magnet.json';
-import exampleFile45 from '../fixtures/manipulate-text-object/manipulate-text-object.json';
-import exampleFile46 from '../fixtures/menu-with-functions-and-text-effects/menu-with-functions-and-text-effects.json';
-import exampleFile47 from '../fixtures/move-camera-to-position/move-camera-to-position.json';
-import exampleFile48 from '../fixtures/move-object-back-and-forth/move-object-back-and-forth.json';
-import exampleFile49 from '../fixtures/move-object-in-circle/move-object-in-circle.json';
-import exampleFile50 from '../fixtures/move-object-toward-position/move-object-toward-position.json';
-import exampleFile51 from '../fixtures/move-object-with-mouse-joint/move-object-with-mouse-joint.json';
-import exampleFile52 from '../fixtures/move-object-with-physics/move-object-with-physics.json';
-import exampleFile53 from '../fixtures/multiplayer-platformer-with-gamepads/multiplayer-platformer-with-gamepads.json';
-import exampleFile54 from '../fixtures/multitouch/multitouch.json';
-import exampleFile55 from '../fixtures/object-gravity/object-gravity.json';
-import exampleFile56 from '../fixtures/object-selection/object-selection.json';
-import exampleFile57 from '../fixtures/objects-timers/objects-timers.json';
-import exampleFile58 from '../fixtures/open-url-in-browser/open-url-in-browser.json';
-import exampleFile59 from '../fixtures/pairs/pairs.json';
-import exampleFile60 from '../fixtures/parallax/parallax.json';
-import exampleFile61 from '../fixtures/parallax-scrolling/parallax-scrolling.json';
-import exampleFile62 from '../fixtures/parse-json-from-api/parse-json-from-api.json';
-import exampleFile63 from '../fixtures/parse-json-string/parse-json-string.json';
-import exampleFile64 from '../fixtures/particles-explosions/particles-explosions.json';
-import exampleFile65 from '../fixtures/particles-various-effects/particles-various-effects.json';
-import exampleFile66 from '../fixtures/pathfinding/pathfinding.json';
-import exampleFile67 from '../fixtures/pathfinding-basics/pathfinding-basics.json';
-import exampleFile68 from '../fixtures/physics/physics.json';
-import exampleFile69 from '../fixtures/physics-joints-demo/physics-joints-demo.json';
-import exampleFile70 from '../fixtures/physics-joints-settings-demo/physics-joints-settings-demo.json';
-import exampleFile71 from '../fixtures/pin-object-to-another/pin-object-to-another.json';
-import exampleFile72 from '../fixtures/pin-object-to-another-multiple-parents/pin-object-to-another-multiple-parents.json';
-import exampleFile73 from '../fixtures/pixel-perfect-platform-game/pixel-perfect-platform-game.json';
-import exampleFile74 from '../fixtures/plane-and-clouds/plane-and-clouds.json';
-import exampleFile75 from '../fixtures/platformer/platformer.json';
-import exampleFile76 from '../fixtures/platformer-double-jump/platformer-double-jump.json';
-import exampleFile77 from '../fixtures/play-music-on-mobile/play-music-on-mobile.json';
-import exampleFile78 from '../fixtures/play-stop-sprite-animation/play-stop-sprite-animation.json';
-import exampleFile79 from '../fixtures/racing-game/racing-game.json';
-import exampleFile80 from '../fixtures/ragdoll/ragdoll.json';
-import exampleFile81 from '../fixtures/rain/rain.json';
-import exampleFile82 from '../fixtures/random-color-picker/random-color-picker.json';
-import exampleFile83 from '../fixtures/rotate-toward-mouse/rotate-toward-mouse.json';
-import exampleFile84 from '../fixtures/rotate-toward-position/rotate-toward-position.json';
-import exampleFile85 from '../fixtures/rotate-with-keypress/rotate-with-keypress.json';
-import exampleFile86 from '../fixtures/save-load/save-load.json';
-import exampleFile87 from '../fixtures/screen-shake/screen-shake.json';
-import exampleFile88 from '../fixtures/shoot-bullet-in-parabola/shoot-bullet-in-parabola.json';
-import exampleFile89 from '../fixtures/shoot-bullets/shoot-bullets.json';
-import exampleFile90 from '../fixtures/shooting-bullets-explanation/shooting-bullets-explanation.json';
-import exampleFile91 from '../fixtures/simple-space-shooter/simple-space-shooter.json';
-import exampleFile92 from '../fixtures/skeletal-animation-demo/skeletal-animation-demo.json';
-import exampleFile93 from '../fixtures/snap-object-to-grid/snap-object-to-grid.json';
-import exampleFile94 from '../fixtures/space-invaders/space-invaders.json';
-import exampleFile95 from '../fixtures/space-shooter/space-shooter.json';
-import exampleFile96 from '../fixtures/splash-screen/splash-screen.json';
-import exampleFile97 from '../fixtures/sprite-fade-in-out/sprite-fade-in-out.json';
-import exampleFile98 from '../fixtures/take-screenshot/take-screenshot.json';
-import exampleFile99 from '../fixtures/text-entry-object/text-entry-object.json';
-import exampleFile100 from '../fixtures/text-fade-in-out/text-fade-in-out.json';
-import exampleFile101 from '../fixtures/text-to-speech/text-to-speech.json';
-import exampleFile102 from '../fixtures/toggle-music-play-sound/toggle-music-play-sound.json';
-import exampleFile103 from '../fixtures/type-on-text-effect/type-on-text-effect.json';
-import exampleFile104 from '../fixtures/video-player/video-player.json';
-import exampleFile105 from '../fixtures/z-depth/z-depth.json';
-import exampleFile106 from '../fixtures/zombie-laser/zombie-laser.json';
-
-// prettier-ignore
-export default {
-  'example://admob': exampleFile0,
-  'example://advanced-shape-based-painter': exampleFile1,
-  'example://animation-speed-scale': exampleFile2,
-  'example://asteroids': exampleFile3,
-  'example://basic-ai-with-pathfinding': exampleFile4,
-  'example://basic-artificial-intelligence': exampleFile5,
-  'example://basic-topdown-car-driving': exampleFile6,
-  'example://betabox-basics-learning-experience': exampleFile7,
-  'example://bomb-the-crate': exampleFile8,
-  'example://bouncing-ball-and-rope': exampleFile9,
-  'example://breakout': exampleFile10,
-  'example://buttons': exampleFile11,
-  'example://car-physics': exampleFile12,
-  'example://center-object-within-another': exampleFile13,
-  'example://change-position-of-object': exampleFile14,
-  'example://change-scale-of-sprites': exampleFile15,
-  'example://change-sprite-animation': exampleFile16,
-  'example://change-sprite-color': exampleFile17,
-  'example://character-selection': exampleFile18,
-  'example://create-object-with-mouseclick': exampleFile19,
-  'example://custom-font': exampleFile20,
-  'example://custom-mouse-pointer': exampleFile21,
-  'example://customize-keys-with-lastpressedkey': exampleFile22,
-  'example://device-orientation-ballgame': exampleFile23,
-  'example://device-orientation-compass': exampleFile24,
-  'example://device-vibration': exampleFile25,
-  'example://dialogue-tree-with-yarn': exampleFile26,
-  'example://downhill-bike-physics-demo': exampleFile27,
-  'example://drag-camera-with-mouse': exampleFile28,
-  'example://drop-collect-items-from-storage': exampleFile29,
-  'example://endless-up-runner': exampleFile30,
-  'example://exit-app': exampleFile31,
-  'example://facebook-instant-game': exampleFile32,
-  'example://filesystem-create-directory': exampleFile33,
-  'example://find-diagonals': exampleFile34,
-  'example://geodash': exampleFile35,
-  'example://health-bar': exampleFile36,
-  'example://infinite-scrolling-background': exampleFile37,
-  'example://inventory-system': exampleFile38,
-  'example://isometric-game': exampleFile39,
-  'example://javascript-blocks-in-platformer': exampleFile40,
-  'example://keyboard-practice': exampleFile41,
-  'example://level-editor': exampleFile42,
-  'example://load-image-from-url': exampleFile43,
-  'example://magnet': exampleFile44,
-  'example://manipulate-text-object': exampleFile45,
-  'example://menu-with-functions-and-text-effects': exampleFile46,
-  'example://move-camera-to-position': exampleFile47,
-  'example://move-object-back-and-forth': exampleFile48,
-  'example://move-object-in-circle': exampleFile49,
-  'example://move-object-toward-position': exampleFile50,
-  'example://move-object-with-mouse-joint': exampleFile51,
-  'example://move-object-with-physics': exampleFile52,
-  'example://multiplayer-platformer-with-gamepads': exampleFile53,
-  'example://multitouch': exampleFile54,
-  'example://object-gravity': exampleFile55,
-  'example://object-selection': exampleFile56,
-  'example://objects-timers': exampleFile57,
-  'example://open-url-in-browser': exampleFile58,
-  'example://pairs': exampleFile59,
-  'example://parallax': exampleFile60,
-  'example://parallax-scrolling': exampleFile61,
-  'example://parse-json-from-api': exampleFile62,
-  'example://parse-json-string': exampleFile63,
-  'example://particles-explosions': exampleFile64,
-  'example://particles-various-effects': exampleFile65,
-  'example://pathfinding': exampleFile66,
-  'example://pathfinding-basics': exampleFile67,
-  'example://physics': exampleFile68,
-  'example://physics-joints-demo': exampleFile69,
-  'example://physics-joints-settings-demo': exampleFile70,
-  'example://pin-object-to-another': exampleFile71,
-  'example://pin-object-to-another-multiple-parents': exampleFile72,
-  'example://pixel-perfect-platform-game': exampleFile73,
-  'example://plane-and-clouds': exampleFile74,
-  'example://platformer': exampleFile75,
-  'example://platformer-double-jump': exampleFile76,
-  'example://play-music-on-mobile': exampleFile77,
-  'example://play-stop-sprite-animation': exampleFile78,
-  'example://racing-game': exampleFile79,
-  'example://ragdoll': exampleFile80,
-  'example://rain': exampleFile81,
-  'example://random-color-picker': exampleFile82,
-  'example://rotate-toward-mouse': exampleFile83,
-  'example://rotate-toward-position': exampleFile84,
-  'example://rotate-with-keypress': exampleFile85,
-  'example://save-load': exampleFile86,
-  'example://screen-shake': exampleFile87,
-  'example://shoot-bullet-in-parabola': exampleFile88,
-  'example://shoot-bullets': exampleFile89,
-  'example://shooting-bullets-explanation': exampleFile90,
-  'example://simple-space-shooter': exampleFile91,
-  'example://skeletal-animation-demo': exampleFile92,
-  'example://snap-object-to-grid': exampleFile93,
-  'example://space-invaders': exampleFile94,
-  'example://space-shooter': exampleFile95,
-  'example://splash-screen': exampleFile96,
-  'example://sprite-fade-in-out': exampleFile97,
-  'example://take-screenshot': exampleFile98,
-  'example://text-entry-object': exampleFile99,
-  'example://text-fade-in-out': exampleFile100,
-  'example://text-to-speech': exampleFile101,
-  'example://toggle-music-play-sound': exampleFile102,
-  'example://type-on-text-effect': exampleFile103,
-  'example://video-player': exampleFile104,
-  'example://z-depth': exampleFile105,
-  'example://zombie-laser': exampleFile106,
-};
