--- newIDE/app/src/ProjectsStorage/BrowserProjectOpener.js
@@ -1,10 +1,0 @@
-import browserExampleFiles from './BrowserExampleFiles';
-
-export default class BrowserProjectOpener {
-  static readInternalFile(url) {
-    if (browserExampleFiles[url])
-      return Promise.resolve(browserExampleFiles[url]);
-
-    return Promise.reject(`Unknown built-in game with URL ${url}`);
-  }
-}
